Know Your City/requirements.txt
pyyaml
requests
pandas
streamlit